Updates state persistence code
Following changes:
1. Makes the state the source of truth in applications
2. Adds error-checking on the initialize API
3. Breaks up the persistor API to multiple pieces to allow for the
   tracking client to use it
4. Adds misc docs/cleans up
